+++ NEWS.md
@@ -58,6 +58,9 @@ This page describes the most important changes in `LazyAlgebra`. The format is b
   `vnorminf` treat complexes as usually done in linear algebra. The only difference is that
   multi-dimensional arguments are considered as *vectors*.
 
+- Extending vectorized methods to other *vector* types shall only require to specialize
+  the *unsafe* version of the methods (the ones with the `unsafe_` prefix).
+
 ### Added
 
 - Most methods can be specialized for specific operator and/or array types by extending